Update main.py
Replaced the old system of loading the games, which was based on previously inputted dictionary in the code. So now instead of it running from a data set in memory, it can now load and save to a JSON file that stores the game dictionary. There was some bugs that came along with that, so had to tidy some extra things.